Rm Tracker construct
